BillingCyclo count

+++ myMoneyApp/myMoney-app/src/main/java/com/example/myMoneyapp/models/CreditModel.java
@@ -22,8 +22,7 @@ public class CreditModel {
     @NotNull
     @Column
     @NumberFormat(style = NumberFormat.Style.NUMBER)
-    private float value;
-    
+    private double value;
     
     public UUID getId() {
         return id;
@@ -41,11 +40,11 @@ public class CreditModel {
         this.name = name;
     }
     
-    public float getValue() {
+    public double getValue() {
         return value;
     }
     
-    public void setValue(float value) {
+    public void setValue(double value) {
         this.value = value;
     }
 }
